// vinta-school-os/src/types/billing.ts
export type PaymentMethod = 'CASH' | 'CCP' | 'BARIDI_MOB'
// ...
export interface MultiPayRequest {
  student_id: string
  items: { group_id: string; amount: number }[]
  total_received: number
  payment_method: PaymentMethod
  pin: string
}